tower

+++ src/csci205_final_project/model/Model.java
@@ -30,7 +30,7 @@ public class Model {
 	return thePlayer;
     }
 
-    public Tower placeTower(int x, int y) {
+    public Tower placeTower(double x, double y) {
         Tower tower = new Tower(x, y);
         return tower;
     }
